requirements.txt
streamlit
langchain
langchain-community
langchain-huggingface
pypdf
chromadb
sentence-transformers
python-dotenv